LPS-80681 Move ExceptionMapper files to exception.mapper package
