+++ modules/apps/apio-architect/apio-architect-api/bnd.bnd
@@ -9,13 +9,13 @@ Export-Package:\
 	com.liferay.apio.architect.alias.routes,\
 	com.liferay.apio.architect.consumer,\
 	com.liferay.apio.architect.consumer.throwable,\
-	com.liferay.apio.architect.converter,\
 	com.liferay.apio.architect.credentials,\
 	com.liferay.apio.architect.date,\
 	com.liferay.apio.architect.documentation,\
 	com.liferay.apio.architect.endpoint,\
 	com.liferay.apio.architect.error,\
 	com.liferay.apio.architect.exception,\
+	com.liferay.apio.architect.exception.mapper,\
 	com.liferay.apio.architect.file,\
 	com.liferay.apio.architect.form,\
 	com.liferay.apio.architect.function,\
