+++ modules/apps/apio-architect/apio-architect-error/src/main/java/com/liferay/apio/architect/error/internal/converter/ClientErrorExceptionMapper.java
@@ -14,8 +14,8 @@
 
 package com.liferay.apio.architect.error.internal.converter;
 
-import com.liferay.apio.architect.converter.ExceptionMapper;
 import com.liferay.apio.architect.error.APIError;
+import com.liferay.apio.architect.exception.mapper.ExceptionMapper;
 
 import javax.ws.rs.ClientErrorException;
 import javax.ws.rs.core.Response;
