+++ modules/apps/apio-architect/apio-architect-error/src/main/java/com/liferay/apio/architect/error/internal/converter/NotAcceptableExceptionMapper.java
@@ -16,8 +16,8 @@ package com.liferay.apio.architect.error.internal.converter;
 
 import static javax.ws.rs.core.Response.Status.NOT_ACCEPTABLE;
 
-import com.liferay.apio.architect.converter.ExceptionMapper;
 import com.liferay.apio.architect.error.APIError;
+import com.liferay.apio.architect.exception.mapper.ExceptionMapper;
 
 import javax.ws.rs.NotAcceptableException;
 import javax.ws.rs.core.Response;
